Сreate property widget factories to apply pattern to flutter fidgets.
