+++ patterns/abstract_factory/tool_panel_factory/widgets/factories_tool_bar.dart
@@ -20,7 +20,8 @@ class FactoriesToolBar extends StatelessWidget {
               ? _buildToolButtons()
               : Padding(
                   padding: EdgeInsets.all(10),
-                  child: CircularProgressIndicator());
+                  child: CircularProgressIndicator(),
+                );
         },
       ),
     );
