+++ patterns/abstract_factory/tool_panel_factory/widgets/property_widgets/bool_property_widget.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+
+import '../../pattern/property.dart';
+import 'factories/property_widget_factory.dart';
+import 'primitive/filed_label.dart';
+
+class BoolPropertyWidgetFactory implements PropertyWidgetFactory {
+  @override
+  Widget createPropertyWidget(Property property) {
+    return BoolPropertyWidget(property: property);
+  }
+
+  @override
+  bool isPropertyContain(Property property) => property.value() is bool;
+}
+
+class BoolPropertyWidget extends StatefulWidget {
+  final Property property;
+
+  const BoolPropertyWidget({Key? key, required this.property})
+      : super(key: key);
+
+  @override
+  _BoolPropertyWidgetState createState() => _BoolPropertyWidgetState();
+}
+
+class _BoolPropertyWidgetState extends State<BoolPropertyWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return FieldLabel(
+      text: widget.property.name,
+      child: Row(
+        children: [
+          Checkbox(
+            value: widget.property.value() as bool,
+            onChanged: (val) {
+              setState(() {
+                widget.property.onChange(val);
+              });
+            },
+          )
+        ],
+      ),
+    );
+  }
+}
