+++ patterns/abstract_factory/tool_panel_factory/widgets/property_widgets/factories/property_widget_factory.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+import '../../../pattern/property.dart';
+
+abstract class PropertyWidgetFactory<T> {
+  Widget createPropertyWidget(Property property);
+
+  bool isPropertyContain(Property value);
+}
